requirements.txt
# Bare minimum
django-cms==3.0.0.beta3
https://github.com/divio/django-cms/archive/3.0.0.beta3.zip

#These dependencies are brought in by django CMS, but if you want to lock-in their version, specify them
Django==1.6.1
django-classy-tags==0.4
South==0.8.1
html5lib==1.0b1
django-mptt==0.5.2
django-sekizai==0.7
six==1.3.0
djangocms-admin-style==0.1.2


#Optional, recommended packages
Pillow==2.0.0
django-filer==0.9.4
cmsplugin-filer==0.9.5
-e git://github.com/etianen/django-reversion.git@f3d20edf96adbb15a1970d3826d6ce8a82021048#egg=django-reversion


#CMS Plugins
djangocms-text-ckeditor

#Zinnia Blog
django-blog-zinnia==0.14
django-tagging==0.3.1

#Zinnia Bootstrap
zinnia-theme-bootstrap
django-app-namespace-template-loader

-e git://github.com/SimonSarazin/django-mailman.git#egg=django-mailman

django-userena==1.2.1
django-social-auth